mapper: Add cursor cursor mapper
This is a wrapper around a SqshMapper that allows the library
to have a moving slice of the archive and dynamicly mmap and
unmap the archive as needed.

For now SqshMapCursor is only a minimal implementation that just
grows the map as needed. In the future it will be extended to
also unmap unused chunks of the archive.