curl: simplify the workflow
before this commit, the curl mapper was updating the mapper
object whenever it was called. With this change, the curl
download function is decoupled from the mapper object. This
way we can use different handlings for the first initial
download to find out the file attributes (size and mtime)
and subsequent downloads check the file attributes.