directory_iterator: fix wrong free, add tests
during one of the last refactorings of the heap constructors,
there was a bug introduced that caused a free of the wrong
pointer. This commit fixes that bug, makes the wrongly freed
pointer const to yield compiler warnings on misuse, and adds
a test to make sure it doesn't happen again.