reader: fix when end offset aligns with block end
This fixes a bug when the data storage wrongfully points to the iterator
when it actually needs to read from the buffer.

A safeguard `assert()` call has been added to the `reader_fill_buffer()`
to detect this error in case of regression. Also a test case has been
added to reproduce this error.