reader: fix handling of when growing a reader
When growing a reader that is in buffered mode, the offset was not
updated correctly. This caused the reader to append overlapping data
twice and therefore corrupting the data.

This commit contains the fix as well as a test case for the metablock
reader. Generalizing the test for the generic reader is left for later.