reader: expect iterator to return bool from _next
Currently the iterator's next function returns an int which
returns >0 on success, 0 on end of iterator and <0 on error.
This commit changes the generic iterator implementation to return
a bool instead of an int. The bool is true if the iterator has
more data and false if the iterator is at the end. The error
code is now passed as a pointer parameter to the next function.