file_iterator: return bool instead of int for _next
This commit changes the return type of sqsh_file_iterator_next() from
int to bool. The return value is now used to indicate whether there is
more data to read from the file iterator. The error code is now stored
in a pointer to an int instead of being returned directly.