easy: fix wrong key length in xattr.
This change fixes a bug in `sqsh_easy_xattr_keys()` where the length of
the key was set to the size of the value. This caused the return value
of `sqsh_easy_xattr_keys()` to contain garbage.

This change also includes a test for sqsh_easy_xattr_keys() and
`sqsh_easy_xattr_get()`.