directory: avoid redundant add/subtract
The amount of entries in a directory fragment is 1-based. That means
that we 0 as the count, we actually have 1 entry left. Before this
change, we would add 1 to the count and subsequently subtract 1 for
each iteration. After this change the count is left as is and we
start subtracting from the count starting from the second iteration.