file_iterator: return error when iterator is not a file
This change fixes the error code that was returned when a file iterator
was initialized with a directory. Before this change, the iterator
treated all file types as regular files and failed when they returned
bogus file addresses. That works, but it is fragile and not the correct
behavior.

This change adds a check to the iterator to ensure that it is initialized
with a file and returns `-SQSH_ERROR_NOT_A_FILE` when it is not.