directory_iterator: replace _inode_number with _inode
This change deprecates sqsh_directory_iterator_inode_number() with a
misleading return type (uint64_t) and replaces it with a new function
sqsh_directory_iterator_inode() with a more appropriate return type (uint32_t).
inodes in squashfs are by definition 32-bit, so there is no need to return a
64-bit. Also the function name is more consistent with the other functions in
libsqsh.