tree_walker: fix _next function.
Up to now the return code of the _next function was always 0, even if
a next entry was found. The user of the walker had no way to check if
the end of the tree was reached. This patch changes the return code
to 1 if there is a next entry and 0 if the end of the directory is
reached as defined in the documentation. It also adds a test for this
case.