directory_iterator: fix potential over-/underflow
This change improves the inode calculation in the directory iterators.
Before this change the inode was checked in check_inode_consistency()
and then calculated naively in directory_iterator_inode().
After this change the inode is precalculated, checked more intensively
and cached in update_inode(). director_iterator_inode() is a simple
getter for the cached value.

This change also includes test cases for the directory iterator that
check the new behavior.