easy: add a function to list directory paths
This change adds `sqsh_easy_directory_list_path()`, a function to list
the contents of a directory as a list of file paths. This enables the
user to get a list of paths that can be directly passed to other easy
functions like `sqsh_easy_file_content()`.

This change also adds a test for the new function.