ls: avoid `newfstatat` syscalls incritical path
When calling `localtime(3)` by default it does a stat to
`/etc/localtime`. It does that repeatedly for any call. As sqsh-ls
has a short lifetime it would be enough (and also reduce the syscall
overhead) to check it once.

Fortunally there's a workaround: These calls can be prevented by setting
the TZ environment variable.[1]

This change sets the TZ variable if unset for `sqsh-ls` which greatly
improves perfomance when listing many files.

[1] https://doc.dovecot.org/configuration_manual/system_calls_optimization/