extract: fix tests when building with `-O2`
When building with optimisations enabled `sqsh__impl_lzo` was inlined
in `extractor.c` this caused the tests to be unable to override this
value to fake an extractor implementation.

This bug also broke the libsqsh-lzo[1] library when building libsqsh
with optimizations enabled.

This change marks `sqsh__impl_lzo` as `volatile` which prevents
inlining.

[1] https://github.com/Gottox/libsqsh-lzo