file: use own getter for the symlink size
This change adds a new getter to retrieve the size of a symlink. Before
this change the file_size getter was used to retrieve the size. This
resulted in situations where the file size could be misinterpreted as a
length for a symlink even though the file wasn't a symlink.