extract: disconnect zlib return value from out value
The return value of the zlib inflate function is not neccessarily in the
right format to be returned directly. While this is not a problem in
itself (libsqsh regards any value smaller than 0 as an error), it is
still a good practice to separate the return value from the out value.